# rest/docs/swagger.yaml
basePath: /
definitions:
...
  api.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  api.PagedResults:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
...
  model.Collection:
    properties:
      c_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      user_id:
        type: integer
      uuid:
        type: string
    type: object
  model.CollectionRef:
    properties:
      collection:
        type: integer
      content:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      path:
        type: string
    type: object
...
host: localhost:3030
info:
  contact:
    email: me@me.com
    name: Me
    url: http://me.com/terms.html
  description: Sample CRUD objects-api for estuary db
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Sample CRUD objects-api for estuary db
  version: "1.0"
paths:
...
  /collectionrefs:
    get:
      consumes:
      - application/json
      description: GetAllCollectionRefs is a handler to get a slice of record(s) from
        collection_refs table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CollectionRef'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of CollectionRefs
      tags:
      - CollectionRefs
    post:
      consumes:
      - application/json
      description: add to add a single record to collection_refs table in the estuary
        database
      parameters:
      - description: Add CollectionRefs
        in: body
        name: CollectionRefs
        required: true
        schema:
          $ref: '#/definitions/model.CollectionRef'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CollectionRef'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to collection_refs table
      tags:
      - CollectionRefs
  /collectionrefs/{argID}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from collection_refs table in the estuary
        database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.CollectionRef'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from collection_refs
      tags:
      - CollectionRefs
    get:
      consumes:
      - application/json
      description: GetCollectionRefs is a function to get a single record from the
        collection_refs table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CollectionRef'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table CollectionRefs by  argID
      tags:
      - CollectionRefs
    put:
      consumes:
      - application/json
      description: Update a single record from collection_refs table in the estuary
        database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      - description: Update CollectionRefs record
        in: body
        name: CollectionRefs
        required: true
        schema:
          $ref: '#/definitions/model.CollectionRef'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CollectionRef'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table collection_refs
      tags:
      - CollectionRefs
  /collections:
    get:
      consumes:
      - application/json
      description: GetAllCollections is a handler to get a slice of record(s) from
        collections table in the estuary database
      parameters:
      - description: page requested (defaults to 0)
        in: query
        name: page
        type: integer
      - description: number of records in a page  (defaults to 20)
        in: query
        name: pagesize
        type: integer
      - description: db sort order column
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.PagedResults'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Collection'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get list of Collections
      tags:
      - Collections
    post:
      consumes:
      - application/json
      description: add to add a single record to collections table in the estuary
        database
      parameters:
      - description: Add Collections
        in: body
        name: Collections
        required: true
        schema:
          $ref: '#/definitions/model.Collection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Add an record to collections table
      tags:
      - Collections
  /collections/{argID}:
    delete:
      consumes:
      - application/json
      description: Delete a single record from collections table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Delete a record from collections
      tags:
      - Collections
    get:
      consumes:
      - application/json
      description: GetCollections is a function to get a single record from the collections
        table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: ErrNotFound, db record for id not found - returns NotFound
            HTTP 404 not found error
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Get record from table Collections by  argID
      tags:
      - Collections
    put:
      consumes:
      - application/json
      description: Update a single record from collections table in the estuary database
      parameters:
      - description: id
        in: path
        name: argID
        required: true
        type: integer
      - description: Update Collections record
        in: body
        name: Collections
        required: true
        schema:
          $ref: '#/definitions/model.Collection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.HTTPError'
      summary: Update an record in table collections
      tags:
      - Collections
...
swagger: "2.0"